+++ webclient/src/components/Common/FormHooks.js
@@ -13,18 +13,6 @@ export function useFormInput(initialValue) {
   };
 }
 
-/*export function useFormInputTextArea(initialValue) {
-  const [value, setValue] = useState(initialValue);
-
-  function handleChange(e) {
-    setValue(e.target.value);
-  }
-
-  return {
-    value,
-    onChange: handleChange
-  };
-}*/
 
 export function useFormCheckbox(initialValue) {
   const [value, setValue] = useState(initialValue);