+++ webclient/src/components/Common/ReactTable/Table.js
@@ -330,8 +330,6 @@ export default function MyTable({ loading, infinite, ...props }) {
         </Row>
         {pagination}
       </Table>
-      <br />
-      <br />
       { /*<JsonTree data={instance} /> */ }
     </div>
   );